Move definitions for template SLB_cin_cam.bt to a single file
