+++ SAGESharp/SAGESharp.csproj
@@ -125,8 +125,7 @@
     <Compile Include="OSI\OSIFile.cs" />
     <Compile Include="OSI\Instruction.cs" />
     <Compile Include="Properties\AssemblyInfo.cs" />
-    <Compile Include="SLB\Cinematic\Camera\Camera.cs" />
-    <Compile Include="SLB\Cinematic\Camera\Frame.cs" />
+    <Compile Include="SLB\Cinematic\Camera.cs" />
     <Compile Include="SLB\Cinematic\Common.cs" />
     <Compile Include="SLB\Common.cs" />
     <Compile Include="SLB\IO\Footer.cs" />
